+++ src/lib/pages/deploy/index.tsx
@@ -6,29 +6,50 @@ import {
   Button,
   Text,
 } from "@chakra-ui/react";
+import { useWallet } from "@cosmos-kit/react";
 import axios from "axios";
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
 
 import { useInternalNavigate, useLCDEndpoint } from "lib/app-provider";
 import { ButtonCard } from "lib/components/ButtonCard";
+import { ConnectWalletAlert } from "lib/components/ConnectWalletAlert";
 import { CustomIcon } from "lib/components/icon";
 import { Stepper } from "lib/components/stepper";
 import WasmPageContainer from "lib/components/WasmPageContainer";
 import { AmpEvent, AmpTrack } from "lib/services/amplitude";
 
-const getUploadAccess = async (endpoint: string) => {
+const getCodeUploadAccess = async (endpoint: string) => {
   const codeParamsUrl = `${endpoint}/cosmwasm/wasm/v1/codes/params`;
   return axios
     .get(codeParamsUrl)
-    .then((res) => res.data.params.code_upload_access.permission);
+    .then((res) => res.data.params.code_upload_access);
+};
+
+const enableUpload = (
+  isPermissionedNetwork: boolean,
+  whitelistedAddresses: string[],
+  address: string | undefined
+) => {
+  let enabled = false;
+  if (!isPermissionedNetwork) {
+    enabled = true;
+  }
+  if (address && isPermissionedNetwork) {
+    enabled = whitelistedAddresses.includes(address);
+  }
+  return enabled;
 };
 
 const Deploy = () => {
   const router = useRouter();
   const navigate = useInternalNavigate();
   const endpoint = useLCDEndpoint();
+  const { address } = useWallet();
   const [isPermissionedNetwork, setIsPermissionedNetwork] = useState(false);
+  const [whitelistedAddresses, setWhitelistedAddresses] = useState<string[]>(
+    []
+  );
 
   useEffect(() => {
     if (router.isReady) AmpTrack(AmpEvent.TO_DEPLOY);
@@ -36,10 +57,11 @@ const Deploy = () => {
 
   useEffect(() => {
     (async () => {
-      const uploadAccess = await getUploadAccess(endpoint);
-      setIsPermissionedNetwork(uploadAccess !== "Everybody");
+      const codeUploadAccess = await getCodeUploadAccess(endpoint);
+      setIsPermissionedNetwork(codeUploadAccess.permission !== "Everybody");
+      setWhitelistedAddresses(codeUploadAccess.addresses);
     })();
-  });
+  }, [endpoint]);
 
   return (
     <WasmPageContainer>
@@ -50,7 +72,11 @@ const Deploy = () => {
       <Heading as="h5" variant="h5" my="48px">
         Select Deploy Option
       </Heading>
-      {isPermissionedNetwork && (
+      <ConnectWalletAlert
+        subtitle="You need to connect wallet to proceed this action"
+        mb={8}
+      />
+      {!enableUpload(isPermissionedNetwork, whitelistedAddresses, address) && (
         <Alert variant="violet" mb="16px" alignItems="flex-start" gap="1">
           <CustomIcon
             name="info-circle-solid"
@@ -58,26 +84,44 @@ const Deploy = () => {
             boxSize="20px"
           />
           <AlertDescription>
-            Uploading new Wasm files on permissioned networks is coming soon to
-            Celatone. Currently, you can upload codes and instantiate contracts
-            on permissionless networks
+            The current network is a permissioned CosmWasm network. Only
+            whitelisted addresses can directly upload Wasm files.
           </AlertDescription>
         </Alert>
       )}
+      {isPermissionedNetwork &&
+        enableUpload(isPermissionedNetwork, whitelistedAddresses, address) && (
+          <Alert mb={8} variant="success">
+            <CustomIcon
+              name="check-circle-solid"
+              color="success.main"
+              boxSize="6"
+              display="flex"
+              alignItems="center"
+            />
+            <AlertDescription mx={4}>
+              Your address is allowed to directly upload Wasm files
+            </AlertDescription>
+          </Alert>
+        )}
       <ButtonCard
         title="Upload new WASM File"
         description={
           isPermissionedNetwork ? (
             <Flex fontSize="14px" gap={1}>
-              <Text color="text.disabled">
-                Currently available on permissionless networks only.
+              <Text
+                color={isPermissionedNetwork ? "text.disabled" : "text.main"}
+              >
+                Available for whitelisted addresses only
               </Text>
             </Flex>
           ) : (
             "Store a new Wasm file on-chain"
           )
         }
-        disabled={isPermissionedNetwork}
+        disabled={
+          !enableUpload(isPermissionedNetwork, whitelistedAddresses, address)
+        }
         onClick={() => navigate({ pathname: "/upload" })}
         mb="16px"
       />
@@ -85,6 +129,7 @@ const Deploy = () => {
         title="Use existing Code IDs"
         description="Input code ID or select from previously stored or saved codes"
         onClick={() => navigate({ pathname: "/instantiate" })}
+        disabled={!address}
       />
       <Flex justify="center" w="100%" mt="32px">
         <Button